Extract generic master-detail components
- master-detail pattern for events created with the use of generic
template
